+++ src/pause.rs
@@ -6,9 +6,9 @@ use leafwing_input_manager::{
 };
 
 use crate::{
-    assets::UiAssets,
     effects::DrawBlinkTimer,
-    state::{is_in_game_state, ForState, GameState},
+    resources::assets::UiAssets,
+    resources::state::{is_in_game_state, ForState, GameState},
 };
 
 #[derive(Actionlike, PartialEq, Eq, Clone, Copy, Hash, Debug, Reflect)]