+++ src/resources/mod.rs
@@ -1,0 +1,3 @@
+pub mod assets;
+pub mod game_time;
+pub mod state;