+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved readability and versatility of assets with bevy_asset_loader
 - Moved menu music loading from main to menu module
 - Moved celestial bodies to new astronomy module
+- Moved assets, game_time and state to resources module
 
 ## [0.0.12] - 2023-10-07